refactor: rename section-header to Pascal case
