+++ app/about/page.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from "react"
 import path from "path"
 import { Hero } from "@/components/Hero"
 import { TextAnimate } from "@/components/magicui/text-animate"
-import { SectionHeader } from "@/components/ui/section-header"
+import { SectionHeader } from "@/components/ui/SectionHeader"
 import { Card, CardContent } from "@/components/ui/Card"
 import { CardWithImage } from "@/components/ui/PeopleCard"
 import { readContentFile } from "@/lib/content-utils"
