+++ app/services/page.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import path from "path"
 import { Hero } from "@/components/Hero"
 import { TextAnimate } from "@/components/magicui/text-animate"
-import { SectionHeader } from "@/components/ui/section-header"
+import { SectionHeader } from "@/components/ui/SectionHeader"
 import { Button } from "@/components/ui/Button"
 import {
   Card,
