+++ components/FeaturedCarousel.tsx
@@ -5,7 +5,7 @@ import { Badge } from "@/components/ui/Badge"
 import { getColorForTag } from "@/lib/utils"
 import { Button } from "@/components/ui/Button"
 import { Card, CardContent } from "@/components/ui/Card"
-import { SectionHeader } from "@/components/ui/section-header"
+import { SectionHeader } from "@/components/ui/SectionHeader"
 import Provident from "@/components/assets/Provident"
 import BrownLogo from "@/components/assets/BrownLogo"
 import { ChevronLeftIcon, ChevronRightIcon, UserIcon } from "lucide-react"
